Client GUI - server GUI - smaller list of connected clients Package reorganization - demo_simulator for tests Final tests - successful
